+++ src/boids/inputs.py
@@ -59,3 +59,9 @@ class InputManager:
 
     def pressed(self, key):
         return key in self._pressed
+
+    def mouse_held(self, key):
+        return self.mouse[key]
+
+    def mouse_pressed(self, key):
+        return key in self._mouse_pressed